+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.1.3+75",
+  "version": "0.1.3+77",
   "private": false,
   "scripts": {
     "dev": "next dev",
