+++ src/interfaces/accounting_account.ts
@@ -1,5 +1,4 @@
 import { ProgressStatus } from '@/constants/account';
-import { isProgressStatus } from '@/lib/utils/type_guard/account';
 
 export interface IAccountingAccount {
   id: number;
@@ -21,14 +20,6 @@ export interface IDetailAccountingAccount {
 export interface IAccountResultStatus {
   resultId: string;
   status: ProgressStatus;
-}
-
-export function isIAccountResultStatus(value: unknown): value is IAccountResultStatus {
-  if (typeof value !== 'object' || value === null) {
-    return false;
-  }
-  const { resultId, status } = value as IAccountResultStatus;
-  return typeof resultId === 'string' && isProgressStatus(status);
 }
 
 export type DetailAccountingAccountOrEmpty = IDetailAccountingAccount | null;
