+++ src/lib/utils/common.ts
@@ -281,3 +281,17 @@ export const countdown = (remainingSeconds: number) => {
 export function eventTypeToVoucherType(eventType: EventType): VoucherType {
   return EVENT_TYPE_TO_VOUCHER_TYPE_MAP[eventType];
 }
+
+// Info Murky (20240505): type guards can input any type and return a boolean
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isStringNumber(value: any): value is string {
+  return typeof value === 'string' && !Number.isNaN(Number(value));
+}
+
+// is {[key: string]: number}
+export function isStringNumberPair(value: unknown): value is { [key: string]: string } {
+  if (typeof value !== 'object' || value === null) {
+    return false;
+  }
+  return Object.values(value).every((v) => typeof v === 'number');
+}
