+++ src/lib/utils/type_guard/account.ts
@@ -1,4 +1,5 @@
 import { EventType, PaymentPeriodType, PaymentStatusType, ProgressStatus, VoucherType } from '@/constants/account';
+import { IAccountResultStatus } from '@/interfaces/accounting_account';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isEventType(data: any): data is EventType {
@@ -23,4 +24,12 @@ export function isPaymentPeriodType(data: any): data is PaymentPeriodType {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isProgressStatus(data: any): data is ProgressStatus {
   return Object.values(ProgressStatus).includes(data as ProgressStatus);
+}
+
+export function isIAccountResultStatus(value: unknown): value is IAccountResultStatus {
+  if (typeof value !== 'object' || value === null) {
+    return false;
+  }
+  const { resultId, status } = value as IAccountResultStatus;
+  return typeof resultId === 'string' && isProgressStatus(status);
 }
